+++ src/commands/fun/SayCommand.js
@@ -9,16 +9,16 @@ module.exports = class SayCommand extends BaseCommand {
   run(client, message, args) {
     const messageToSay = args.join(" ");
     const sayEmbed = new Discord.MessageEmbed()
-    .setTitle(`${messageToSay}`)
-    .setFooter(message.author.tag ,message.author.displayAvatarURL())
-    .setColor("#4daf8")
-    .setTimestamp();
+      .setTitle(`${messageToSay}`)
+      .setFooter(message.author.tag, message.author.displayAvatarURL())
+      .setColor("#4daf8")
+      .setTimestamp();
     try {
       message.channel.send(sayEmbed);
     } catch (err) {
       console.log(err);
-      message.channel.send('I am unable to send that message.");')
+      message.channel.send('I am unable to send that message.')
     }
-    message.delete();
+    if (message.guild.me.hasPermission("MANAGE_MESSAGES")) { message.delete() };
   }
 }
