+++ src/commands/information/SocialCommand.js
@@ -25,6 +25,6 @@ module.exports = class SocialCommand extends BaseCommand {
     message.channel.send(twitchEmbed).catch(err => console.log(err));
     message.channel.send(discordEmbed).catch(err => console.log(err));
     message.channel.send(inviteEmbed).catch(err => console.log(err));
-    message.delete();
+    if (message.guild.me.hasPermission("MANAGE_MESSAGES")) { message.delete() };
   }
 }
